Changed quote attributed name extraction method
Used alternate method that does not rely on book.animateEntities for extracting phrase names. This was mostly done due to the changes added in the quote attribution method that required a quote-attributed name to be valid only if it came from a phrase put into animateEntities. The  improvements I made to the quote-attribution program stood in contradiciton with this, as the name I extracted for quote-attribution did not always stay in animateEntities. I could have added that particular 'phrase' containing the name to animateEntities instead, but I decided to subvert the requirement itself as I did not want to meddle with the code for extracting phrases unnecessarily. The code I wrote would add phrases which the phrase-generating code did not deem as legitimate to add. As my program does not need this requirement anyway (I don't know if requirement is the correct word - I saw the code few weeks ago, and all I can say is that all the quote-attributed names happen to be from phrases in animateEntities - I don't remember whether the name dictates whether the phrase is added or the other way round), I just skipped it and used an alternate method that requires the printHTML option to be processed first as can be seen in my changes to BookNLP.java